:bug: Fix(CHAT): ChannelInterceptor 설정
