+++ src/main/java/com/runto/domain/chat/config/WebSocketSecurityInterceptor.java
@@ -1,0 +1,62 @@
+package com.runto.domain.chat.config;
+
+import com.runto.global.security.detail.CustomUserDetails;
+import com.runto.global.security.dto.UserDetailsDTO;
+import com.runto.global.security.util.JWTUtil;
+import lombok.RequiredArgsConstructor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.messaging.Message;
+import org.springframework.messaging.MessageChannel;
+import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
+import org.springframework.messaging.support.ChannelInterceptor;
+import org.springframework.messaging.support.MessageHeaderAccessor;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Component;
+
+import java.util.List;
+
+@RequiredArgsConstructor
+@Component
+public class WebSocketSecurityInterceptor implements ChannelInterceptor {
+    private static final Logger log = LoggerFactory.getLogger(WebSocketConfig.class);
+    private final JWTUtil jwtUtil;
+
+    @Override
+    public Message<?> preSend(Message<?> message, MessageChannel channel) {
+
+        StompHeaderAccessor headerAccessor = MessageHeaderAccessor.getAccessor(message, StompHeaderAccessor.class);
+        String authHeader = headerAccessor.getFirstNativeHeader("Authorization");
+
+        if (authHeader != null && authHeader.startsWith("Bearer ")) {
+            String token = authHeader.substring(7);
+            try {
+                Long userId = jwtUtil.getId(token);  // JWT에서 사용자 정보 추출
+                log.info("InboundChannel userId = {}",userId);
+                String username = jwtUtil.getUsername(token);
+                log.info("InboundChannel username = {}",username);
+                String role = jwtUtil.getRole(token);
+
+                // 인증된 사용자 정보 설정
+                CustomUserDetails userDetails = new CustomUserDetails(new UserDetailsDTO(userId,null,null,null,username,null,role));
+                log.info("InboundChannel userDetails userId = {}",userDetails.getUserId());
+                Authentication authenticationToken = new UsernamePasswordAuthenticationToken(userDetails, null, List.of(new SimpleGrantedAuthority(role)));
+                SecurityContextHolder.getContext().setAuthentication(authenticationToken);
+                log.info("InboundChannel Context Authentication = {}",SecurityContextHolder.getContext().getAuthentication());
+                SecurityContextHolder.setStrategyName(SecurityContextHolder.MODE_INHERITABLETHREADLOCAL);
+
+            } catch (Exception e) {
+                throw new RuntimeException("Invalid token: " + e.getMessage());
+            }
+        }
+        return message;
+    }
+
+    @Override
+    public void afterSendCompletion(Message<?> message, MessageChannel channel, boolean sent, Exception ex) {
+        SecurityContextHolder.clearContext();
+    }
+}
